added function get_fiedler so it is more portable and can potentially be used standalone somewhere else. Added functionality for the genetic algorithm, though the current changes make gaopt routine to probably not run standalone since function calls are slightly different.

+++ include/IntegralMatrix.h
@@ -162,7 +162,8 @@ class OneElectronArray
     {
       if (bin) {
         dumpFile.seekg (0, ios::end);
-	int size = dumpFile.tellg();
+	//int size = dumpFile.tellg();
+	double size = dumpFile.tellg();
         dumpFile.seekg (0, ios::beg);
         FORTINT nmo = rhf ? static_cast<int>(2*sqrt(size / (sizeof(double)))) : static_cast<int>(sqrt(size / (sizeof(double))));
 	ReSize(nmo);
